Changed to PlaySpec and some cleanup
